% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg.R
\name{check_bioc}
\alias{check_bioc}
\title{check_bioc}
\usage{
check_bioc(pkg = "BiocInstaller")
}
\arguments{
\item{pkg}{package name}
}
\value{
list
}
\description{
check latest release version of bioconductor package
}
\examples{
\dontrun{
library(rvcheck)
check_bioc('ggtree')
}
}
\author{
Guangchuang Yu
}

