% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg.R
\name{check_cran}
\alias{check_cran}
\title{check_cran}
\usage{
check_cran(pkg)
}
\arguments{
\item{pkg}{package name}
}
\description{
check latest release version of cran package
}
\examples{
\dontrun{
library(rvcheck)
check_cran('emojifont')
}
}
\author{
Guangchuang Yu
}

