% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg.R
\name{check_github}
\alias{check_github}
\title{check_github}
\usage{
check_github(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
list
}
\description{
check latest github version of R package
}
\author{
Guangchuang Yu
}

