% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg.R
\name{check_gitlab}
\alias{check_gitlab}
\title{check_gitlab}
\usage{
check_gitlab(pkg)
}
\arguments{
\item{pkg}{package name}
}
\value{
list
}
\description{
check latest gitlab version of R package
}
\author{
Guangchuang Yu
}
