% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_aes_var}
\alias{get_aes_var}
\title{get_aes_var}
\usage{
get_aes_var(mapping, var)
}
\arguments{
\item{mapping}{aes mapping}

\item{var}{variable}
}
\value{
mapped var
}
\description{
extract aes mapping, compatible with ggplot2 < 2.3.0 & > 2.3.0
}
\author{
guangchuang yu
}
