% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.installed}
\alias{is.installed}
\title{is.installed}
\usage{
is.installed(packages)
}
\arguments{
\item{packages}{selected packagtes}
}
\value{
logical vector
}
\description{
check whether packages were installed
}
\details{
check whether selected packages were installed
}
\author{
Guangchuang Yu
}
