% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_fun_from_pkg}
\alias{is.installed}
\alias{o}
\alias{read.cb}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{yulab.utils}{\code{\link[yulab.utils]{get_fun_from_pkg}}, \code{\link[yulab.utils]{is.installed}}, \code{\link[yulab.utils]{o}}, \code{\link[yulab.utils]{read.cb}}}
}}

